streamlit/requirements.txt
streamlit
requests
soundfile
numpy
matplotlib
streamlit-audiorecorder

